Fix Filesystem tests
