+++ tests/toolchain/CsvCompoundToolchainTest.php
@@ -47,8 +47,8 @@ class CsvCompoundToolchainTest extends MikTestBase
      */
     protected function setUp()
     {
-        parent::setUp();
         $this->path_to_temp_dir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . "mik_csv_compound_temp_dir";
+        parent::setUp();
         $this->path_to_output_dir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . "mik_csv_compound_output_dir";
         @mkdir($this->path_to_output_dir);
         $this->path_to_log = $this->path_to_temp_dir . DIRECTORY_SEPARATOR . "mik.log";
