+++ tests/toolchain/CsvSingleFileToolchainTest.php
@@ -32,8 +32,8 @@ class CsvSingleFileToolchainTest extends MikTestBase
      */
     protected function setUp()
     {
-        parent::setUp();
         $this->path_to_temp_dir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . "mik_csv_fetcher_temp_dir";
+        parent::setUp();
         $this->path_to_output_dir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . "mik_csv_single_file_output_dir";
         $this->path_to_log = $this->path_to_temp_dir . DIRECTORY_SEPARATOR . "mik.log";
         $this->path_to_input_validator_log = $this->path_to_temp_dir . DIRECTORY_SEPARATOR . "input_validator.log";
